typo fixes

+++ tests/u128_mul_div_test.rs
@@ -46,8 +46,8 @@ mod u128_mul_div_test {
 
     #[simplex::test]
     fn u128_test_mul_div_128_product_is_u128(context: simplex::TestContext) -> anyhow::Result<()> {
-        let a = rand::thread_rng().gen_range(0..=u16::MAX) as u128;
-        let b = rand::thread_rng().gen_range(0..=u16::MAX) as u128;
+        let a = rand::thread_rng().gen_range(0..=u64::MAX) as u128;
+        let b = rand::thread_rng().gen_range(0..=u64::MAX) as u128;
         let c = rand::thread_rng().gen_range(1..=u128::MAX);
 
         let res = a * b / c;
